added permissions for groups and users

+++ users/views.py
@@ -6,9 +6,12 @@ from django.contrib import messages
 from django.http import HttpResponse
 from django.views.decorators.csrf import csrf_exempt
 import json
+from users.decorators import init_check, init_check_access , init_check_profile
+
 
 # Create your views here.
 @login_required(login_url="login/")
+@init_check_profile
 def profile(request):
     if request.method == "POST":
         user = request.user
@@ -30,7 +33,14 @@ def profile(request):
 
 
 @login_required(login_url="login/")
+@init_check
 def add_group(request):
+    curr_user = request.user
+    curr_user_group= curr_user.groups.all()
+    curr_user_permissions = Access.objects.get(group = curr_user_group[0].id)
+    if 'Manage Groups' not in curr_user_permissions.menu : 
+        return HttpResponse('you are not authorized here')
+    
     if request.method == "POST":
         groupName = request.POST['groupName']
         permissions = request.POST.getlist('permissions')
@@ -49,35 +59,15 @@ def add_group(request):
     }
     return render(request , 'groups/add_group.html' , context)
 
-
 @login_required(login_url="login/")
-@csrf_exempt
-def manage_access(request):
-    if request.method == 'POST':
-        a = json.loads(request.body)
-        for i in range(0, a["length"]):
-            user = User.objects.get(email = a[str(i)]["email"])
-            group = Group.objects.filter(name=a[str(i)]["group"])
-            user.groups.add(group[0])
-            return redirect('manage_access')
-          
-    users = User.objects.all()
-    groups = Group.objects.all()
-    a = []
-    for user in users:
-        if not user.groups.exists():
-            a.append(user)
-            
-
-    context = {
-        'users' : a ,
-        'groups': groups,
-    }
-    return render(request , 'users/manage_access.html' , context)
-
-
-@login_required(login_url="login/")
+@init_check
 def manage_permission(request):    
+    curr_user = request.user
+    curr_user_group= curr_user.groups.all()
+    curr_user_permissions = Access.objects.get(group = curr_user_group[0].id)
+    if 'Manage Groups' not in curr_user_permissions.menu : 
+        return HttpResponse('you are not authorized here')
+    
     accesses = Access.objects.all()
     context = {
         'accesses':accesses
@@ -86,7 +76,14 @@ def manage_permission(request):
 
 
 @login_required(login_url="login/")
+@init_check
 def manage_pre_update(request , id):
+    curr_user = request.user
+    curr_user_group= curr_user.groups.all()
+    curr_user_permissions = Access.objects.get(group = curr_user_group[0].id)
+    if 'Manage Groups' not in curr_user_permissions.menu : 
+        return HttpResponse('you are not authorized here')
+    
     if request.method == "POST":
         permissions = request.POST.getlist('permissions')
         print(permissions)
@@ -110,7 +107,14 @@ def manage_pre_update(request , id):
 
 
 @login_required(login_url="login/")
+@init_check
 def manage_pre_delete(request , id):
+    curr_user = request.user
+    curr_user_group= curr_user.groups.all()
+    curr_user_permissions = Access.objects.get(group = curr_user_group[0].id)
+    if 'Manage Groups' not in curr_user_permissions.menu : 
+        return HttpResponse('you are not authorized here')
+    
     acc = Access.objects.get(id = id)
     grp = Group.objects.get(name = acc.group.name)
     grp.delete()
@@ -121,7 +125,14 @@ def manage_pre_delete(request , id):
 
 @login_required(login_url="login/")
 @csrf_exempt
+@init_check
 def manage_user_permission(request):
+    curr_user = request.user
+    curr_user_group= curr_user.groups.all()
+    curr_user_permissions = Access.objects.get(group = curr_user_group[0].id)
+    if 'Manage Users' not in curr_user_permissions.menu : 
+        return HttpResponse('you are not authorized here')
+    
     if request.method == 'POST':
         a = json.loads(request.body)
         for i in range(0, a["length"]):
@@ -146,6 +157,42 @@ def manage_user_permission(request):
         'groups': groups,
     }
     return render(request , 'users/update_access.html' , context)
+
+
+@login_required(login_url="login/")
+@csrf_exempt
+def manage_access(request):
+    curr_user = request.user
+    if not curr_user.groups.exists():
+        return HttpResponse('Wait till we authorize you .')
+    if not request.user.related_profiles.all():
+        return redirect('profile')
+    curr_user_group= curr_user.groups.all()
+    curr_user_permissions = Access.objects.get(group = curr_user_group[0].id)
+    if 'Manage Users' not in curr_user_permissions.menu : 
+        return HttpResponse('you are not authorized here')
+    
+    if request.method == 'POST':
+        a = json.loads(request.body)
+        for i in range(0, a["length"]):
+            user = User.objects.get(email = a[str(i)]["email"])
+            group = Group.objects.filter(name=a[str(i)]["group"])
+            user.groups.add(group[0])
+            return redirect('manage_access')
+          
+    users = User.objects.all()
+    groups = Group.objects.all()
+    a = []
+    for user in users:
+        if not user.groups.exists():
+            a.append(user)
+            
+
+    context = {
+        'users' : a ,
+        'groups': groups,
+    }
+    return render(request , 'users/manage_access.html' , context)
